Always send a SIGTERM to children when using 'kill' or 'reset'

+++ daemon/linux/process_helper.rs
@@ -3,23 +3,36 @@ use ::nix::{
     sys::signal::{self, Signal},
     unistd::Pid,
 };
-use procfs::process::all_processes;
+use procfs::process::{all_processes, Process};
 
-/// A small helper to send a signal to all direct child processes of a specific task.
-/// `pid` is the process id of the main task id. I.e the parent pid of the child processes.
-pub fn send_signal_to_children(pid: i32, signal: Signal) {
-    let children = all_processes()
+/// Get all children of a specific process
+pub fn get_children(pid: i32) -> Vec<Process> {
+    all_processes()
         .unwrap()
         .into_iter()
-        .filter(|process| process.stat.ppid == pid);
+        .filter(|process| process.stat.ppid == pid)
+        .collect()
+}
 
-    for child in children {
-        let pid = Pid::from_raw(child.pid);
+/// Send a signal to a list of processes
+pub fn send_signal_to_processes(processes: Vec<Process>, signal: Signal) {
+    for process in processes {
+        // Process is no longer alive, skip this.
+        if !process.is_alive() {
+            continue
+        }
+
+        let pid = Pid::from_raw(process.pid);
         if let Err(error) = signal::kill(pid, signal) {
             warn!(
                 "Failed send signal {:?} to Pid {}: {:?}",
-                signal, child.pid, error
+                signal, process.pid, error
             );
         }
     }
+}
+
+/// A small helper that sends a signal to all children of a specific process by id.
+pub fn send_signal_to_children(pid: i32, signal: Signal) {
+    send_signal_to_processes(get_children(pid), signal);
 }
